+++ crates/box/src/lib.rs
@@ -211,4 +211,3 @@ fn sequence_runner(
 
     Ok(())
 }
-