+++ crates/pretty_text/src/effects/appearance/scramble.rs
@@ -16,7 +16,7 @@ use super::Appeared;
 
 pub(super) fn plugin(app: &mut App) {
     app.add_systems(
-        Update,
+        PostUpdate,
         (mark_effect_glyphs::<Scramble, ComputeScramble>, scramble)
             .chain()
             .in_set(PrettyEffectSystems),