+++ crates/pretty_text/src/effects/appearance/spread.rs
@@ -13,7 +13,7 @@ use super::Appeared;
 
 pub(super) fn plugin(app: &mut App) {
     app.add_systems(
-        Update,
+        PostUpdate,
         (mark_effect_glyphs::<Spread, ComputeSpread>, spread)
             .chain()
             .in_set(PrettyEffectSystems),