+++ crates/pretty_text/src/effects/behavior/bounce.rs
@@ -11,7 +11,7 @@ use crate::glyph::{GlyphIndex, GlyphVertices, SpanGlyphOf, VertexMask};
 
 pub(super) fn plugin(app: &mut App) {
     app.add_systems(
-        Update,
+        PostUpdate,
         (mark_effect_glyphs::<Bounce, BounceTimeline>, shake)
             .chain()
             .in_set(PrettyEffectSystems),