+++ crates/pretty_text/src/effects/behavior/fade.rs
@@ -8,7 +8,7 @@ use crate::glyph::{GlyphIndex, GlyphVertices, SpanGlyphOf};
 
 pub(super) fn plugin(app: &mut App) {
     app.add_systems(
-        Update,
+        PostUpdate,
         (mark_effect_glyphs::<Fade, ComputeFade>, fade)
             .chain()
             .in_set(PrettyEffectSystems),