+++ crates/pretty_text/src/effects/behavior/shake.rs
@@ -12,7 +12,7 @@ use crate::glyph::{GlyphVertices, SpanGlyphOf, VertexMask};
 
 pub(super) fn plugin(app: &mut App) {
     app.add_systems(
-        Update,
+        PostUpdate,
         (mark_effect_glyphs::<Shake, ShakeOffset>, shake)
             .chain()
             .in_set(PrettyEffectSystems),