+++ crates/pretty_text/src/effects/behavior/wobble.rs
@@ -12,7 +12,7 @@ use crate::glyph::{GlyphIndex, GlyphVertices, SpanGlyphOf, VertexMask};
 
 pub(super) fn plugin(app: &mut App) {
     app.add_systems(
-        Update,
+        PostUpdate,
         (mark_effect_glyphs::<Wobble, ComputeWobble>, wobble)
             .chain()
             .in_set(PrettyEffectSystems),